+++ milkman-rest/src/main/java/milkman/ui/plugin/rest/domain/RestRequestContainer.java
@@ -1,5 +1,7 @@
 package milkman.ui.plugin.rest.domain;
 
+import java.util.UUID;
+
 import org.apache.commons.lang3.StringUtils;
 
 import lombok.AllArgsConstructor;